backend/requirements-dev.txt
-r requirements.txt
ruff==0.8.*
pytest==8.*
pytest-django==4.9.*
pytest-cov==6.*
factory-boy==3.3.*
ipython==8.*
django-debug-toolbar==4.4.*
